Launcher should show the R48 version properly
So now you can easily identify all your Lost R48 Builds, like:
"R48 v2.0-versiontest-2"
"v1.6-4-mouse2"
etc.

+++ releaser/releaser-pre.sh
@@ -9,6 +9,11 @@
 # Supply with the version name and version code.
 # This sets up the release metadata, then runs the Maven compile/package.
 
+if [ "$#" -ne 3 ]; then
+ echo "releaser-pre.sh RELEASEID ANDROIDVERSIONCODE DEVFLAG"
+ exit 1
+fi
+
 # Write release metadata
 rm -rf metadata/src || exit
 mkdir -p metadata/src/main/resources/assets || exit
@@ -18,7 +23,12 @@ VERSIONFILE="metadata/src/main/resources/assets/version.txt"
 cp ../CREDITS.txt metadata/src/main/resources/assets/ || exit
 cp ../COPYING.txt metadata/src/main/resources/assets/ || exit
 # The date is represented with the last commit's date.
-echo "R48 $1 (AVC $2), last commit:" `git show-ref HEAD` `git log | grep Date | head -n 1` >> $VERSIONFILE || exit
+if [ "$3" = 1 ]; then
+ echo "R48 $1 [debug]" >> $VERSIONFILE || exit
+else
+ echo "R48 $1" >> $VERSIONFILE || exit
+fi
+echo "AVC $2, last commit:" `git show-ref HEAD` `git log | grep Date | head -n 1` >> $VERSIONFILE || exit
 # Write in the boring details
 echo "gabien-app-r48 - Editing program for various formats" >> $VERSIONFILE || exit
 echo "Written starting in 2016 by contributors (see CREDITS.txt)" >> $VERSIONFILE || exit
